callguard-ai/mcp_server/requirements.txt
fastmcp>=2.0.0
scalekit-sdk-python>=2.6.1
python-dotenv>=1.1.1
